+++ src/lib/Core/CHANGELOG.md
@@ -1,4 +1,8 @@
-﻿# 3.1.4
+﻿# 3.1.5
+* Renamed Matrix4.empty and Vector4.create to Matrix4/Vector4.zero.
+* Added more linear algebra functions.
+
+# 3.1.4
 * Added "allAtoms" query.
 
 # 3.1.3